128-bits gpoupOrder representation + big merge with Leo

+++ k-Factorization/TripleSys/Include/SRGsupport.h
@@ -9,7 +9,7 @@ public:
 	SrgSummary(const std::string* logFolder, int iMode, const std::string* extFolder);
 	void copyFiles(const std::string& inp, const std::string& out);
 	void deleteFile(const std::string& tmp);
-	void outSRG_info(int v, const SRGParam* graphParam, t_graphType graphType, int rank3, size_t grOrder, int srcGroupSize, int srcGroups, int srcAut);
+	void outSRG_info(int v, const SRGParam* graphParam, t_graphType graphType, int rank3, UInt grOrder, int srcGroupSize, int srcGroups, int srcAut);
 private:
 	std::string outFile;
 	std::string outFile2;
